requirements.txt
# Core dependencies
streamlit
streamlit-nested-layout
streamlit-extras
# Data processing and visualization
pandas
openpyxl
plotly
statsmodels
# Database
psycopg2-binary
sqlalchemy
# API and configuration
python-dotenv
openai
jsonschema
# statistical models
statsmodels
